{{cookiecutter.project_name}}/requirements.txt
aiohttp==0.21.6
python-memcached==1.58
requests
click
python-dateutil==2.5.3
cchardet==1.0.0
aiohttp-debugtoolbar==0.1.1
aiohttp-session==0.5.0
aiohttp-jinja2==0.7.0
aioredis==0.2.7
motor==0.6.2
cryptography
commentjson
pymongo==3.3.1
MongoAlchemy
bcrypt
validate_email
prometheus_client
rq
redis
jinja2
Flask
Flask-Admin
Flask-Login>=0.3.0
flask-appbuilder
flask-mongoengine
Flask-Session
